ROTATIONAL FINAL DRAFT + update to dynamics script
finally got the rotator code to the point where it will probably work - need to test it on the machine in astrolab.

Now defines a double for xdot and loads that into DotNet Numerics rk45 integrator to spit out new state vector.

Updated dynamics model to now grab intial parameters from unity scene and intitialize all states, as well as call rotate3 on fixedupdate.
Added basic button-press checks to be able to alter external torque vector for the purposes of testing.

Unless something doesn't work, this should be the last update to RotateOrion3.